skills/self-learning-agent/requirements.txt
python-dotenv
requests
python-wordpress-xmlrpc
